requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
numpy==1.26.4
scikit-learn==1.4.0
tensorflow-cpu==2.16.1  # Updated to 2.16.1 which fixed several serialization issues
gunicorn==21.2.0
Werkzeug==3.0.1
requests==2.32.3
protobuf>=4.21.1,<5.0.0  
h5py==3.10.0
psutil==5.9.5
mysql-connector-python==8.2.0
google-cloud-storage==2.10.0
# python-dotenv==1.0.0
